+++ doc/source/changelog.md
@@ -26,6 +26,7 @@ This branch is the last supporting Godot 3.
     - `VoxelGeneratorGraph`: fixed memory leaks when the graph contains resources
     - `VoxelTerrain`: fixed `Condition "mesh_block == nullptr" is true` which could happen in some conditions
     - `VoxelTool`: `raycast` locking up if you send a Vector3 containing NaN
+    - `VoxelToolLodTerrain`: fix inconsistent result with integer `do_sphere` radius
     - `VoxelInstancer`: fix instances not refreshing when an item is modified and the mesh block size is 32
     - `VoxelInstancer`: fix crash when removing an item from the library while an instancer node is using it
     - `VoxelInstancer`: fix errors when removing scene instances